Convert tests to functional form
