+++ Project.toml
@@ -17,9 +17,22 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 [compat]
 BlockDiagonals = "0.1"
 ChainRulesCore = "1"
+HiGHS = "1"
+Ipopt = "1"
 IterativeSolvers = "0.9"
-JuMP = "0.23, 1"
+JuMP = "1"
 LazyArrays = "0.21, 0.22, 1"
 MathOptInterface = "1"
 MathOptSetDistances = "0.2"
+SCS = "1"
 julia = "1.6"
+
+[extras]
+DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
+HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
+Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
+SCS = "c946c3f1-0d1f-5ce8-9dea-7daa1f7e2d13"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["DelimitedFiles", "HiGHS", "Ipopt", "SCS", "Test"]
