--- test/Project.toml
@@ -1,19 +1,0 @@
-[deps]
-ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
-JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
-Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
-HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
-Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
-MLDatasets = "eb30cadb-4394-5ae3-aed4-317e484a6458"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-SCS = "c946c3f1-0d1f-5ce8-9dea-7daa1f7e2d13"
-Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-Ipopt = "1.0.2"
