+++ test/moi_wrapper.jl
@@ -3,38 +3,47 @@
 # Use of this source code is governed by an MIT-style license that can be found
 # in the LICENSE.md file or at https://opensource.org/licenses/MIT.
 
+module TestMOIWrapper
+
 using Test
-import DiffOpt
-import MathOptInterface as MOI
 import DelimitedFiles
-import Ipopt
+import DiffOpt
 import HiGHS
+import Ipopt
+import MathOptInterface as MOI
 import SCS
 
-const VAF = MOI.VectorAffineFunction{Float64}
-_vaf(c::Vector{Float64}) = VAF(MOI.ScalarAffineTerm{Float64}[], c)
+const ATOL = 2e-4
+const RTOL = 2e-4
 
-@testset "MOI Unit" begin
-    function test_runtests()
-        model = DiffOpt.diff_optimizer(HiGHS.Optimizer)
-        # `Variable.ZerosBridge` makes dual needed by some tests fail.
-        MOI.Bridges.remove_bridge(
-            model.optimizer.optimizer,
-            MOI.Bridges.Variable.ZerosBridge{Float64},
-        )
-        MOI.set(model, MOI.Silent(), true)
-        config = MOI.Test.Config(; atol = 1e-7)
-        MOI.Test.runtests(model, config), return
+function runtests()
+    for name in names(@__MODULE__; all = true)
+        if startswith("$name", "test_")
+            @testset "$(name)" begin
+                getfield(@__MODULE__, name)()
+            end
+        end
     end
-    test_runtests()
 end
 
-@testset "Testing forward on trivial QP" begin
+include(joinpath(@__DIR__, "utils.jl"))
+
+function test_moi_test_runtests()
+    model = DiffOpt.diff_optimizer(HiGHS.Optimizer)
+    # `Variable.ZerosBridge` makes dual needed by some tests fail.
+    MOI.Bridges.remove_bridge(
+        model.optimizer.optimizer,
+        MOI.Bridges.Variable.ZerosBridge{Float64},
+    )
+    MOI.set(model, MOI.Silent(), true)
+    config = MOI.Test.Config(; atol = 1e-7)
+    MOI.Test.runtests(model, config)
+    return
+end
+
+function test_forward_on_trivial_QP()
     # using example on https://osqp.org/docs/examples/setup-and-solve.html
-    Q = [
-        4.0 1.0
-        1.0 2.0
-    ]
+    Q = [4.0 1.0; 1.0 2.0]
     q = [1.0, 1.0]
     G = [
         1.0 1.0
@@ -59,9 +68,10 @@ end
         # Expected solutions
         z = [0.3, 0.7],
     )
+    return
 end
 
-@testset "Differentiating trivial QP 1" begin
+function test_differentiating_trivial_qp_1()
     Q = [
         4.0 1.0
         1.0 2.0
@@ -89,6 +99,7 @@ end
         z = [-0.25; -0.75],
         dhb = ones(1),
     )
+    return
 end
 
 # @testset "Differentiating a non-convex QP" begin
@@ -96,10 +107,10 @@ end
 #     q = [1.0; 1.0]
 #     G = [1.0 1.0;]
 #     h = [-1.0;]
-
+#
 #     model = DiffOpt.diff_optimizer(Ipopt.Optimizer)
 #     x = MOI.add_variables(model, 2)
-
+#
 #     # define objective
 #     quad_terms = MOI.ScalarQuadraticTerm{Float64}[]
 #     for i in 1:2
@@ -110,7 +121,7 @@ end
 #             )
 #         end
 #     end
-
+#
 #     objective_function = MOI.ScalarQuadraticFunction(
 #                             quad_terms,
 #                             MOI.ScalarAffineTerm.(q, x),
@@ -118,21 +129,20 @@ end
 #                         )
 #     MOI.set(model, MOI.ObjectiveFunction{MOI.ScalarQuadraticFunction{Float64}}(), objective_function)
 #     MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
-
+#
 #     # add constraint
 #     MOI.add_constraint(
 #         model,
 #         MOI.ScalarAffineFunction(MOI.ScalarAffineTerm.(G[1, :], x), 0.0),
 #         MOI.LessThan(h[1]),
 #     )
-
+#
 #     @test_throws ErrorException MOI.optimize!(model) # should break
 # end
 
-@testset "Differentiating QP with inequality and equality constraints" begin
+function test_differentiating_qp_with_inequality_and_equality_constraints()
     # refered from: https://www.mathworks.com/help/optim/ug/quadprog.html#d120e113424
     # Find equivalent qpth program here - https://github.com/AKS1996/jump-gsoc-2020/blob/master/DiffOpt_tests_4_py.ipynb
-
     Q = [
         1.0 -1.0 1.0
         -1.0 2.0 -2.0
@@ -150,7 +160,6 @@ end
     h = [1.0, 1.0, 1.0, 0.0, 0.0, 0.0]
     A = [1.0 1.0 1.0;]
     b = [0.5]
-
     qp_test_with_solutions(
         Ipopt.Optimizer;
         Q = Q,
@@ -175,18 +184,18 @@ end
         dAb = [0.0 -0.5 0.0],
         dbb = [1.0],
     )
+    return
 end
 
 # refered from https://github.com/jump-dev/MathOptInterface.jl/blob/master/src/Test/contquadratic.jl#L3
 # Find equivalent CVXPYLayers and QPTH code here:
 #               https://github.com/AKS1996/jump-gsoc-2020/blob/master/DiffOpt_tests_1_py.ipynb
-@testset "Differentiating MOI examples 1" begin
+function test_differentiating_moi_examples_1()
     # homogeneous quadratic objective
     # Min x^2 + xy + y^2 + yz + z^2
     # st  x + 2y + 3z >= 4 (c1)
     #     x +  y      >= 1 (c2)
     #     x, y, z \in R
-
     Q = [
         2.0 1.0 0.0
         1.0 2.0 1.0
@@ -198,20 +207,17 @@ end
         -1.0 -1.0 0.0
     ]
     h = [-4.0, -1.0]
-
     dQ = [
         -0.12244895 0.01530609 -0.11224488
         0.01530609 0.09183674 0.07653058
         -0.11224488 0.07653058 -0.06122449
     ]
     dq = [-0.2142857, 0.21428567, -0.07142857]
-
     dG = [
         0.05102692 0.30612244 0.25510856
         0.06120519 0.36734693 0.30610315
     ]
     dh = [-0.35714284; -0.4285714]
-
     qp_test_with_solutions(
         Ipopt.Optimizer;
         Q = Q,
@@ -229,17 +235,17 @@ end
         dGb = dG,
         dhb = dh,
     )
+    return
 end
 
 # refered from https://github.com/jump-dev/MathOptInterface.jl/blob/master/src/Test/contquadratic.jl#L3
 # Find equivalent CVXPYLayers and QPTH code here:
 #               https://github.com/AKS1996/jump-gsoc-2020/blob/master/DiffOpt_tests_2_py.ipynb
-@testset "Differentiating MOI examples 2" begin
+function test_differentiating_moi_examples_2()
     # non-homogeneous quadratic objective
     #    minimize 2 x^2 + y^2 + xy + x + y
     #       s.t.  x, y >= 0
     #             x + y = 1
-
     Q = [
         4 1.0
         1 2
@@ -252,7 +258,6 @@ end
     h = [0, 0.0]
     A = [1 1.0]
     b = [1.0]
-
     dQ = [
         -0.05 -0.05
         -0.05 0.15
@@ -262,7 +267,6 @@ end
     dh = zeros(2)
     dA = [0.375 -1.075]
     db = [0.7]
-
     qp_test_with_solutions(
         Ipopt.Optimizer;
         Q = Q,
@@ -297,59 +301,37 @@ end
         dνb = zeros(1),
         ∇νb = [-0.7],
     )
+    return
 end
 
-@testset "Differentiating non trivial convex QP MOI" begin
+function test_differentiating_non_trivial_convex_qp_moi()
     nz = 10
     nineq_le = 25
     neq = 10
-
     # read matrices from files
     names = ["P", "q", "G", "h", "A", "b"]
     matrices = []
-
     for name in names
-        push!(
-            matrices,
-            DelimitedFiles.readdlm(
-                joinpath(
-                    dirname(dirname(pathof(DiffOpt))),
-                    "test",
-                    "data",
-                    name * ".txt",
-                ),
-                ' ',
-                Float64,
-                '\n',
-            ),
-        )
+        filename = joinpath(@__DIR__, "data", "$name.txt")
+        push!(matrices, DelimitedFiles.readdlm(filename, ' ', Float64, '\n'))
     end
-
     Q, q, G, h, A, b = matrices
     q = vec(q)
     h = vec(h)
     b = vec(b)
-
     # read gradients from files
     names = ["dP", "dq", "dG", "dh", "dA", "db"]
     grads_actual = []
-
     for name in names
+        filename = joinpath(@__DIR__, "data", "$name.txt")
         push!(
             grads_actual,
-            DelimitedFiles.readdlm(
-                joinpath(@__DIR__, "data", name * ".txt"),
-                ' ',
-                Float64,
-                '\n',
-            ),
+            DelimitedFiles.readdlm(filename, ' ', Float64, '\n'),
         )
     end
-
     dqb = vec(grads_actual[2])
     dhb = vec(grads_actual[4])
     dbb = vec(grads_actual[6])
-
     qp_test(
         Ipopt.Optimizer,
         true,
@@ -375,9 +357,10 @@ end
         atol = 1e-3, # The values in `data` seems to have low accuracy
         rtol = 1e-3, # The values in `data` seems to have low accuracy
     )
+    return
 end
 
-@testset "Differentiating LP; checking gradients for non-active contraints" begin
+function test_differentiating_LP_checking_gradients_for_non_active_contraints()
     # Issue #40 from Gurobi.jl
     # min  x
     # s.t. x >= 0
@@ -394,9 +377,10 @@ end
         dGb = [0.0, 3.0],
         dhb = [0.0, -1.0],
     )
+    return
 end
 
-@testset "Differentiating a simple LP with GreaterThan constraint" begin
+function test_differentiating_a_simple_LP_with_GreaterThan_constraint()
     # this is canonically same as above test
     # min  x
     # s.t. x >= 3
@@ -412,11 +396,11 @@ end
         dGb = [3.0],
         dhb = [-1.0],
     )
+    return
 end
 
-@testset "Differentiating LP; checking gradients for non-active contraints" begin
+function test_differentiating_lp()
     # refered from - https://en.wikipedia.org/wiki/Simplex_algorithm#Example
-
     # max 2x + 3y + 4z
     # s.t. 3x+2y+z <= 10
     #      2x+5y+3z <= 15
@@ -446,9 +430,10 @@ end
         ],
         dhb = [0.0, 1 / 3, -1 / 3, 2 / 3, 0.0],
     )
+    return
 end
 
-@testset "Differentiating LP with variable bounds" begin
+function test_differentiating_LP_with_variable_bounds()
     # max 2x + 3y + 4z
     # s.t. 3x+2y+z <= 10
     #      2x+5y+3z <= 15
@@ -488,9 +473,10 @@ end
         ],
         dhb = [0.0, 1 / 3, -1 / 3, 2 / 3, 0.0, 0.0, 0.0, 0.0],
     )
+    return
 end
 
-@testset "Differentiating LP with variable bounds 2" begin
+function test_differentiating_LP_with_variable_bounds_2()
     nz = 3
     qp_test_with_solutions(
         HiGHS.Optimizer;
@@ -518,19 +504,20 @@ end
         dAb = [0.0 0.0 -5 / 3],
         dbb = [1 / 3],
     )
+    return
 end
 
-@testset "Differentiating LP with SAF, SV with LE, GE constraints" begin
-    """
-        max  2x + 3y + 4z
-        s.t. 3x+2y+z <= 10
-             2x+5y+3z <= 15
-             x ≤ 3
-             0 ≤ y ≤ 2
-             6 ≥ z ≥ -1
-             x, y, z >= 0
-        variant of previous test with same solution
-    """
+"""
+max  2x + 3y + 4z
+s.t. 3x+2y+z <= 10
+     2x+5y+3z <= 15
+     x ≤ 3
+     0 ≤ y ≤ 2
+     6 ≥ z ≥ -1
+     x, y, z >= 0
+variant of previous test with same solution
+"""
+function test_differentiating_lp_with_saf_with_le_ge_constraints()
     nz = 3
     qp_test_with_solutions(
         HiGHS.Optimizer;
@@ -562,16 +549,16 @@ end
         ],
         dhb = [0.0, 1 / 3, -1 / 3, 2 / 3, 0.0, 0.0, 0.0, 0.0, 0.0],
     )
+    return
 end
 
 # TODO: split file here
 # above is QP Back
 # below is conic forw
 
-function test_simple_socp(eq_vec::Bool)
+function _test_simple_socp(eq_vec::Bool)
     # referred from _soc2test, https://github.com/jump-dev/MathOptInterface.jl/blob/master/src/Test/contconic.jl#L1355
     # find equivalent diffcp python program here: https://github.com/AKS1996/jump-gsoc-2020/blob/master/diffcp_socp_1_py.ipynb
-
     # Problem SOC2
     # min  x
     # s.t. y ≥ 1/√2
@@ -581,18 +568,15 @@ function test_simple_socp(eq_vec::Bool)
     # s.t.  -1/√2 + y ∈ R₊
     #        1 - t ∈ {0}
     #      (t,x,y) ∈ SOC₃
-
     model = DiffOpt.diff_optimizer(SCS.Optimizer)
     MOI.set(model, MOI.Silent(), true)
     x, y, t = MOI.add_variables(model, 3)
-
     MOI.set(
         model,
         MOI.ObjectiveFunction{MOI.ScalarAffineFunction{Float64}}(),
         1.0x,
     )
     MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
-
     if eq_vec
         ceq = MOI.add_constraint(
             model,
@@ -612,9 +596,7 @@ function test_simple_socp(eq_vec::Bool)
         MOI.Utilities.vectorize([1.0t, 1.0x, 1.0y]),
         MOI.SecondOrderCone(3),
     )
-
     MOI.optimize!(model)
-
     if eq_vec
         MOI.set(
             model,
@@ -637,9 +619,7 @@ function test_simple_socp(eq_vec::Bool)
         csoc,
         MOI.Utilities.operate(vcat, Float64, 1.0 * t, 0.0, 0.0),
     )
-
     DiffOpt.forward_differentiate!(model)
-
     # these matrices are benchmarked with the output generated by diffcp
     # refer the python file mentioned above to get equivalent python source code
     @test model.diff.model.x ≈ [-1 / √2; 1 / √2; 1.0] atol = ATOL rtol = RTOL
@@ -654,7 +634,6 @@ function test_simple_socp(eq_vec::Bool)
         @test model.diff.model.y ≈ [1.0, √2, 1.0, -1.0, √2] atol = ATOL rtol =
             RTOL
     end
-
     dx = [1.12132144; 1 / √2; 1 / √2]
     for (i, vi) in enumerate([x, y, t])
         @test dx[i] ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), vi) atol =
@@ -663,13 +642,12 @@ function test_simple_socp(eq_vec::Bool)
     # @test dx ≈ [1.12132144; 1/√2; 1/√2] atol=ATOL rtol=RTOL
     # @test ds ≈ [0.0; 0.0; -2.92893438e-01;  1.12132144e+00; 7.07106999e-01]  atol=ATOL rtol=RTOL
     # @test dy ≈ [2.4142175; 5.00000557; 3.8284315; √2; -4.00000495] atol=ATOL rtol=RTOL
+    return
 end
 
-@testset "Differentiating simple SOCP" begin
-    for eq_vec in [false, true]
-        test_simple_socp(eq_vec)
-    end
-end
+test_differentiating_simple_SOCP_vector() = _test_simple_socp(true)
+
+test_differentiating_simple_SOCP_scalar() = _test_simple_socp(false)
 
 # refered from _psd0test, https://github.com/jump-dev/MathOptInterface.jl/blob/master/src/Test/contconic.jl#L3919
 # find equivalent diffcp program here: https://github.com/AKS1996/jump-gsoc-2020/blob/master/diffcp_sdp_1_py.ipynb
@@ -683,17 +661,16 @@ end
 #     ⎛ 1   1 ⎞
 # X = ⎜       ⎟           y = 2
 #     ⎝ 1   1 ⎠
-function simple_psd(solver)
-    model = DiffOpt.diff_optimizer(solver)
+function test_simple_psd()
+    model = DiffOpt.diff_optimizer(SCS.Optimizer)
     MOI.set(model, MOI.Silent(), true)
     X = MOI.add_variables(model, 3)
     vov = MOI.VectorOfVariables(X)
     cX = MOI.add_constraint(
         model,
-        VAF(vov),
+        MOI.VectorAffineFunction{Float64}(vov),
         MOI.PositiveSemidefiniteConeTriangle(2),
     )
-
     c = MOI.add_constraint(
         model,
         MOI.VectorAffineFunction(
@@ -702,18 +679,14 @@ function simple_psd(solver)
         ),
         MOI.Zeros(1),
     )
-
     MOI.set(
         model,
         MOI.ObjectiveFunction{MOI.ScalarAffineFunction{Float64}}(),
         1.0 * X[1] + 1.0 * X[end],
     )
     MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
-
     MOI.optimize!(model)
-
     x = MOI.get(model, MOI.VariablePrimal(), X)
-
     cone_types = unique([
         S for (F, S) in
         MOI.get(model.optimizer, MOI.ListOfConstraintTypesPresent())
@@ -722,28 +695,29 @@ function simple_psd(solver)
     cones = conic_form.constraints.sets
     DiffOpt.set_set_types(cones, cone_types)
     index_map = MOI.copy_to(conic_form, model)
-
     # s = DiffOpt.map_rows((ci, r) -> MOI.get(model.optimizer, MOI.ConstraintPrimal(), ci), model.optimizer, cones, index_map, DiffOpt.Flattened{Float64}())
     # y = DiffOpt.map_rows((ci, r) -> MOI.get(model.optimizer, MOI.ConstraintDual(), ci), model.optimizer, cones, index_map, DiffOpt.Flattened{Float64}())
-
     @test x ≈ ones(3) atol = ATOL rtol = RTOL
     # @test s ≈ [0.0; ones(3)] atol=ATOL rtol=RTOL
     # @test y ≈ [2.0, 1.0, -1.0, 1.0]  atol=ATOL rtol=RTOL
-
     # test1: changing the constant in `c`, i.e. changing value of X[2]
-    MOI.set(model, DiffOpt.ForwardConstraintFunction(), c, _vaf([1.0]))
-
+    MOI.set(
+        model,
+        DiffOpt.ForwardConstraintFunction(),
+        c,
+        MOI.VectorAffineFunction{Float64}(
+            MOI.ScalarAffineTerm{Float64}[],
+            [1.0],
+        ),
+    )
     DiffOpt.forward_differentiate!(model)
-
     dx = -ones(3)
     for (i, vi) in enumerate(X)
         @test dx[i] ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), vi) atol =
             ATOL rtol = RTOL
     end
-
     # @test dx ≈ -ones(3) atol=ATOL rtol=RTOL  # will change the value of other 2 variables
     # @test ds[2:4] ≈ -ones(3)  atol=ATOL rtol=RTOL  # will affect PSD constraint too
-
     # test2: changing X[1], X[3] but keeping the objective (their sum) same
     MOI.set(
         model,
@@ -755,22 +729,17 @@ function simple_psd(solver)
         ),
     )
     MOI.set(model, DiffOpt.ForwardObjectiveFunction(), -1.0X[1] + 1.0X[3])
-
     DiffOpt.forward_differentiate!(model)
-
     # @test dx ≈ [1.0, 0.0, -1.0] atol=ATOL rtol=RTOL  # note: no effect on X[2]
     dx = [1.0, 0.0, -1.0]
     for (i, vi) in enumerate(X)
         @test dx[i] ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), vi) atol =
             ATOL rtol = RTOL
     end
+    return
 end
 
-@testset "Differentiating simple PSD program" begin
-    simple_psd(SCS.Optimizer)
-end
-
-@testset "Differentiating conic with PSD and SOC constraints" begin
+function test_differentiating_conic_with_PSD_and_SOC_constraints()
     # similar to _psd1test, https://github.com/jump-dev/MathOptInterface.jl/blob/master/src/Test/contconic.jl#L4054
     # find equivalent diffcp example here - https://github.com/AKS1996/jump-gsoc-2020/blob/master/diffcp_sdp_2_py.ipynb
 
@@ -798,10 +767,8 @@ end
     #      |    y2    y2 y1+y2 |
     #
     #      (1-y1, -y2, -y2) in C^3_q
-
     model = DiffOpt.diff_optimizer(SCS.Optimizer)
     MOI.set(model, MOI.Silent(), true)
-
     δ = √(1 + (3 * √2 + 2) * √(-116 * √2 + 166) / 14) / 2
     ε = √((1 - 2 * (√2 - 1) * δ^2) / (2 - √2))
     y2 = 1 - ε * δ
@@ -811,22 +778,19 @@ end
     x2 = ((3 - 2obj) * (2 + k^2) - 4) / (4 * (2 + k^2) - 4 * √2)
     α = √(3 - 2obj - 4x2) / 2
     β = k * α
-
     X = MOI.add_variables(model, 6)
     x = MOI.add_variables(model, 3)
-
     vov = MOI.VectorOfVariables(X)
     cX = MOI.add_constraint(
         model,
-        VAF(vov),
+        MOI.VectorAffineFunction{Float64}(vov),
         MOI.PositiveSemidefiniteConeTriangle(3),
     )
     cx = MOI.add_constraint(
         model,
-        VAF(MOI.VectorOfVariables(x)),
+        MOI.VectorAffineFunction{Float64}(MOI.VectorOfVariables(x)),
         MOI.SecondOrderCone(3),
     )
-
     c1 = MOI.add_constraint(
         model,
         MOI.VectorAffineFunction(
@@ -855,7 +819,6 @@ end
         ),
         MOI.Zeros(1),
     )
-
     # this is a useless constraint - refer the tests below
     # even if we comment this, it won't affect the optimal values
     c_extra = MOI.add_constraint(
@@ -866,7 +829,6 @@ end
         ),
         MOI.Nonnegatives(1),
     )
-
     objXidx = [1:3; 5:6]
     objXcoefs = 2 * ones(5)
     MOI.set(
@@ -878,12 +840,9 @@ end
         ),
     )
     MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
-
     MOI.optimize!(model)
-
     _x = MOI.get(model, MOI.VariablePrimal(), x)
     _X = MOI.get(model, MOI.VariablePrimal(), X)
-
     cone_types = unique([
         S for (F, S) in
         MOI.get(model.optimizer, MOI.ListOfConstraintTypesPresent())
@@ -892,10 +851,8 @@ end
     cones = conic_form.constraints.sets
     DiffOpt.set_set_types(cones, cone_types)
     index_map = MOI.copy_to(conic_form, model)
-
     # s = DiffOpt.map_rows((ci, r) -> MOI.get(model.optimizer, MOI.ConstraintPrimal(), ci), model.optimizer, cones, index_map, DiffOpt.Flattened{Float64}())
     # y = DiffOpt.map_rows((ci, r) -> MOI.get(model.optimizer, MOI.ConstraintDual(), ci), model.optimizer, cones, index_map, DiffOpt.Flattened{Float64}())
-
     @test _X ≈ [
         0.21725121
         -0.25996907
@@ -905,7 +862,6 @@ end
         0.21725121
     ] atol = ATOL rtol = RTOL
     @test _x ≈ [0.2544097; 0.17989425; 0.17989425] atol = ATOL rtol = RTOL
-
     # @test x ≈ [ 0.21725121; -0.25996907;  0.31108582;  0.21725009; -0.25996907;  0.21725121;
     #             0.2544097;   0.17989425;  0.17989425] atol=ATOL rtol=RTOL
     # @test s ≈ [
@@ -922,12 +878,17 @@ end
     #     0.455242, -0.321905, -0.321905,
     #     1.13334, 0.678095, 1.13334, -0.321905, 0.678095, 1.13334,
     # ]  atol=ATOL rtol=RTOL
-
     # test c_extra
-    MOI.set(model, DiffOpt.ForwardConstraintFunction(), c_extra, _vaf([1.0]))
-
+    MOI.set(
+        model,
+        DiffOpt.ForwardConstraintFunction(),
+        c_extra,
+        MOI.VectorAffineFunction{Float64}(
+            MOI.ScalarAffineTerm{Float64}[],
+            [1.0],
+        ),
+    )
     DiffOpt.forward_differentiate!(model)
-
     # a small change in the constant in c_extra should not affect any other variable or constraint other than c_extra itself
     for (i, vi) in enumerate(X)
         @test 0.0 ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), vi) atol =
@@ -941,20 +902,17 @@ end
     # @test dy ≈ zeros(12) atol=0.012
     # @test [ds[1:2]; ds[4:end]] ≈ zeros(11) atol=1e-2
     # @test ds[3] ≈ 1.0 atol=1e-2   # except c_extra itself
+    return
 end
 
-@testset "Differentiating conic with PSD and POS constraints" begin
+function test_differentiating_conic_with_PSD_and_POS_constraints()
     # refer psdt2test, https://github.com/jump-dev/MathOptInterface.jl/blob/master/src/Test/contconic.jl#L4306
     # find equivalent diffcp program here - https://github.com/AKS1996/jump-gsoc-2020/blob/master/diffcp_sdp_3_py.ipynb
-
     model = DiffOpt.diff_optimizer(SCS.Optimizer)
     MOI.set(model, MOI.Silent(), true)
-
     x = MOI.add_variables(model, 7)
     @test MOI.get(model, MOI.NumberOfVariables()) == 7
-
     η = 10.0
-
     c1 = MOI.add_constraint(
         model,
         MOI.VectorAffineFunction(
@@ -1017,16 +975,13 @@ end
         ),
         MOI.Zeros(1),
     )
-
     MOI.set(
         model,
         MOI.ObjectiveFunction{MOI.ScalarAffineFunction{Float64}}(),
         MOI.ScalarAffineFunction([MOI.ScalarAffineTerm(1.0, x[7])], 0.0),
     )
     MOI.set(model, MOI.ObjectiveSense(), MOI.MAX_SENSE)
-
     MOI.optimize!(model)
-
     # dc = ones(7)
     MOI.set(
         model,
@@ -1059,9 +1014,7 @@ end
         c4,
         MOI.Utilities.vectorize(ones(1, 7) * x + ones(1)),
     )
-
     DiffOpt.forward_differentiate!(model)
-
     @test model.diff.model.x ≈
           [20 / 3.0, 0.0, 10 / 3.0, 0.0, 0.0, 0.0, 1.90192379] atol = ATOL rtol =
         RTOL
@@ -1091,10 +1044,8 @@ end
         0.408248,
         0.78867513,
     ] atol = ATOL rtol = RTOL
-
     atol = 0.3
     rtol = 0.01
-
     # compare these with https://github.com/AKS1996/jump-gsoc-2020/blob/master/diffcp_sdp_3_py.ipynb
     # results are not exactly as: 1. there is some residual error   2. diffcp results are SCS specific, hence scaled
     dx = [-39.6066, 10.8953, -14.9189, 10.9054, 10.883, 10.9118, -21.7508]
@@ -1104,13 +1055,11 @@ end
     end
     # @test dy ≈ [0.0, -3.56905, 0.0, -0.380035, 0.0, -0.41398, -0.385321, -0.00743119, -0.644986, -0.550542, -2.36765] atol=atol rtol=rtol
     # @test ds ≈ [0.0, 0.0, -50.4973, 0.0, -25.8066, 0.0, 0.0, 0.0, -7.96528, -1.62968, -2.18925] atol=atol rtol=rtol
-
     # TODO: future example, how to differentiate wrt a specific constraint/variable, refer QPLib article for more
     dA = zeros(11, 7)
     dA[3:8, 1:6] = Matrix{Float64}(LinearAlgebra.I, 6, 6)  # differentiating only wrt POS constraint c2
     db = zeros(11)
     dc = zeros(7)
-
     # db = zeros(11)
     # dA = zeros(11, 7)
     # dA[3:8, 1:6] = Matrix{Float64}(LinearAlgebra.I, 6, 6)  # differentiating only wrt POS constraint c2
@@ -1118,7 +1067,10 @@ end
         model,
         DiffOpt.ForwardConstraintFunction(),
         c1,
-        MOI.Utilities.zero_with_output_dimension(VAF, 1),
+        MOI.Utilities.zero_with_output_dimension(
+            MOI.VectorAffineFunction{Float64},
+            1,
+        ),
     )
     MOI.set(
         model,
@@ -1130,30 +1082,33 @@ end
         model,
         DiffOpt.ForwardConstraintFunction(),
         c3,
-        MOI.Utilities.zero_with_output_dimension(VAF, 3),
+        MOI.Utilities.zero_with_output_dimension(
+            MOI.VectorAffineFunction{Float64},
+            3,
+        ),
     )
     MOI.set(
         model,
         DiffOpt.ForwardConstraintFunction(),
         c4,
-        MOI.Utilities.zero_with_output_dimension(VAF, 1),
+        MOI.Utilities.zero_with_output_dimension(
+            MOI.VectorAffineFunction{Float64},
+            1,
+        ),
     )
-
     DiffOpt.forward_differentiate!(model)
-
     # for (i, vi) in enumerate(X)
     #     @test 0.0 ≈ MOI.get(model,
     #         DiffOpt.ForwardVariablePrimal(), vi) atol=1e-2 rtol=RTOL
     # end
-
     # TODO add a test here, probably on duals
-
     # # note that there's no change in the PSD slack values or dual optimas
     # @test dy ≈ [0.0, 0.0, 0.0, 0.125978, 0.0, 0.142644, 0.142641, 0.0127401, 0.0, 0.0, 0.0] atol=atol rtol=RTOL
     # @test ds ≈ [0.0, 0.0, -6.66672, 0.0, -3.33336, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0] atol=atol rtol=RTOL
+    return
 end
 
-@testset "Differentiating a simple PSD" begin
+function test_differentiating_a_simple_psd()
     # refer _psd3test, https://github.com/jump-dev/MathOptInterface.jl/blob/master/src/Test/contconic.jl#L4484
     # find equivalent diffcp program here - https://github.com/AKS1996/jump-gsoc-2020/blob/master/diffcp_sdp_0_py.ipynb
 
@@ -1161,58 +1116,44 @@ end
     # s.t. [x 1 1]
     #      [1 x 1] ⪰ 0
     #      [1 1 x]
-
     model = DiffOpt.diff_optimizer(SCS.Optimizer)
     MOI.set(model, MOI.Silent(), true)
-
     x = MOI.add_variable(model)
-
-    func = MOI.Utilities.operate(
-        vcat,
-        Float64,
-        x,
-        one(Float64),
-        x,
-        one(Float64),
-        one(Float64),
-        x,
-    )
-
+    func = MOI.Utilities.operate(vcat, Float64, x, 1.0, x, 1.0, 1.0, x)
     # do not confuse this constraint with the matrix `c` in the conic form (of the matrices A, b, c)
     c = MOI.add_constraint(model, func, MOI.PositiveSemidefiniteConeTriangle(3))
-
     MOI.set(
         model,
         MOI.ObjectiveFunction{MOI.ScalarAffineFunction{Float64}}(),
         MOI.ScalarAffineFunction(MOI.ScalarAffineTerm.(1.0, [x]), 0.0),
     )
     MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
-
     MOI.optimize!(model)
-
     # SCS/Mosek specific
     # @test s' ≈ [1.         1.41421356 1.41421356 1.         1.41421356 1.        ] atol=ATOL rtol=RTOL
     # @test y' ≈ [ 0.33333333 -0.23570226 -0.23570226  0.33333333 -0.23570226  0.33333333]  atol=ATOL rtol=RTOL
-
     # dA = zeros(6, 1)
     # db = ones(6)
-    MOI.set(model, DiffOpt.ForwardConstraintFunction(), c, _vaf(ones(6)))
+    MOI.set(
+        model,
+        DiffOpt.ForwardConstraintFunction(),
+        c,
+        MOI.VectorAffineFunction{Float64}(
+            MOI.ScalarAffineTerm{Float64}[],
+            ones(6),
+        ),
+    )
     # dc = zeros(1)
-
     DiffOpt.forward_differentiate!(model)
-
     @test model.diff.model.x ≈ [1.0] atol = 10ATOL rtol = 10RTOL
     @test model.diff.model.s ≈ ones(6) atol = ATOL rtol = RTOL
     @test model.diff.model.y ≈ [1 / 3, -1 / 6, 1 / 3, -1 / 6, -1 / 6, 1 / 3] atol =
         ATOL rtol = RTOL
-
     @test -0.5 ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), x) atol = 1e-2 rtol =
         RTOL
-
     # @test dx ≈ [-0.5] atol=ATOL rtol=RTOL
     # @test dy ≈ zeros(6) atol=ATOL rtol=RTOL
     # @test ds ≈ [0.5, 1.0, 0.5, 1.0, 1.0, 0.5] atol=ATOL rtol=RTOL
-
     # test 2
     dA = zeros(6, 1)
     db = zeros(6)
@@ -1220,33 +1161,29 @@ end
         model,
         DiffOpt.ForwardConstraintFunction(),
         c,
-        MOI.Utilities.zero_with_output_dimension(VAF, 6),
+        MOI.Utilities.zero_with_output_dimension(
+            MOI.VectorAffineFunction{Float64},
+            6,
+        ),
     )
     MOI.set(model, DiffOpt.ForwardObjectiveFunction(), 1.0 * x)
-
     DiffOpt.forward_differentiate!(model)
-
     @test 0.0 ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), x) atol = 1e-2 rtol =
         RTOL
-
     # @test dx ≈ zeros(1) atol=ATOL rtol=RTOL
     # @test dy ≈ [0.333333, -0.333333, 0.333333, -0.333333, -0.333333, 0.333333] atol=ATOL rtol=RTOL
     # @test ds ≈ zeros(6) atol=ATOL rtol=RTOL
+    return
 end
 
-@testset "Verifying cache after differentiating a QP" begin
-    Q = [
-        4.0 1.0
-        1.0 2.0
-    ]
+function test_verifying_cache_after_differentiating_a_qp()
+    Q = [4.0 1.0; 1.0 2.0]
     q = [1.0, 1.0]
     G = [1.0 1.0]
     h = [-1.0]
-
     model = DiffOpt.diff_optimizer(SCS.Optimizer)
     MOI.set(model, MOI.Silent(), true)
     x = MOI.add_variables(model, 2)
-
     # define objective
     quad_terms = MOI.ScalarQuadraticTerm{Float64}[]
     for i in 1:2
@@ -1265,7 +1202,6 @@ end
         MOI.ObjectiveFunction{MOI.ScalarQuadraticFunction{Float64}}(),
         objective_function,
     )
-
     # add constraint
     c = MOI.add_constraint(
         model,
@@ -1273,33 +1209,24 @@ end
         MOI.LessThan(h[1]),
     )
     MOI.optimize!(model)
-
     x_sol = MOI.get(model, MOI.VariablePrimal(), x)
     @test x_sol ≈ [-0.25, -0.75] atol = ATOL rtol = RTOL
-
     MOI.set.(model, DiffOpt.ReverseVariablePrimal(), x, ones(2))
-
     @test model.diff === nothing
     DiffOpt.reverse_differentiate!(model)
-
     grad_wrt_h =
         MOI.constant(MOI.get(model, DiffOpt.ReverseConstraintFunction(), c))
     @test grad_wrt_h ≈ -1.0 atol = 2ATOL rtol = RTOL
-
     # adding two variables invalidates the cache
     y = MOI.add_variables(model, 2)
     MOI.delete(model, y)
-
     @test model.diff === nothing
     MOI.optimize!(model)
-
     MOI.set.(model, DiffOpt.ReverseVariablePrimal(), x, ones(2))
     DiffOpt.reverse_differentiate!(model)
-
     grad_wrt_h =
         MOI.constant(MOI.get(model, DiffOpt.ReverseConstraintFunction(), c))
     @test grad_wrt_h ≈ -1.0 atol = 2ATOL rtol = RTOL
-
     # adding single variable invalidates the cache
     y0 = MOI.add_variable(model)
     @test model.diff === nothing
@@ -1308,17 +1235,14 @@ end
         MOI.ScalarAffineFunction([MOI.ScalarAffineTerm(1.0, y0)], 0.0),
         MOI.EqualTo(42.0),
     )
-
     MOI.optimize!(model)
     @test model.diff === nothing
-
     MOI.set.(model, DiffOpt.ReverseVariablePrimal(), x, ones(2))
     DiffOpt.reverse_differentiate!(model)
     grad_wrt_h =
         MOI.constant(MOI.get(model, DiffOpt.ReverseConstraintFunction(), c))
     @test grad_wrt_h ≈ -1.0 atol = 5e-3 rtol = RTOL
     @test model.diff.model.gradient_cache isa DiffOpt.QuadraticProgram.Cache
-
     # adding constraint invalidates the cache
     c2 = MOI.add_constraint(
         model,
@@ -1327,7 +1251,6 @@ end
     )
     @test model.diff === nothing
     MOI.optimize!(model)
-
     MOI.set.(model, DiffOpt.ReverseVariablePrimal(), x, ones(2))
     DiffOpt.reverse_differentiate!(model)
     grad_wrt_h =
@@ -1338,18 +1261,15 @@ end
         MOI.constant(MOI.get(model, DiffOpt.ReverseConstraintFunction(), c2))
     @test grad_wrt_h ≈ 0.0 atol = 5e-3 rtol = RTOL
     @test model.diff.model.gradient_cache isa DiffOpt.QuadraticProgram.Cache
+    return
 end
 
-@testset "Verifying cache on a PSD" begin
+function test_verifying_cache_on_a_psd()
     model = DiffOpt.diff_optimizer(SCS.Optimizer)
     MOI.set(model, MOI.Silent(), true)
-
     x = MOI.add_variable(model)
-
     func = MOI.Utilities.operate(vcat, Float64, x, 1.0, x, 1.0, 1.0, x)
-
     c = MOI.add_constraint(model, func, MOI.PositiveSemidefiniteConeTriangle(3))
-
     MOI.set(
         model,
         MOI.ObjectiveFunction{MOI.ScalarAffineFunction{Float64}}(),
@@ -1359,35 +1279,290 @@ end
     @test model.diff === nothing
     MOI.optimize!(model)
     @test model.diff === nothing
-
     dA = zeros(6, 1)
     db = ones(6)
-    MOI.set(model, DiffOpt.ForwardConstraintFunction(), c, _vaf(ones(6)))
+    MOI.set(
+        model,
+        DiffOpt.ForwardConstraintFunction(),
+        c,
+        MOI.VectorAffineFunction{Float64}(
+            MOI.ScalarAffineTerm{Float64}[],
+            ones(6),
+        ),
+    )
     dc = zeros(1)
-
     DiffOpt.forward_differentiate!(model)
-
     @test model.diff.model.x ≈ [1.0] atol = 10ATOL rtol = 10RTOL
     @test model.diff.model.s ≈ ones(6) atol = ATOL rtol = RTOL
     @test model.diff.model.y ≈ [1 / 3, -1 / 6, 1 / 3, -1 / 6, -1 / 6, 1 / 3] atol =
         ATOL rtol = RTOL
-
     @test -0.5 ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), x) atol = 1e-2 rtol =
         RTOL
-
     @test model.diff.model.gradient_cache isa DiffOpt.ConicProgram.Cache
-
     DiffOpt.forward_differentiate!(model)
-
     @test -0.5 ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), x) atol = 1e-2 rtol =
         RTOL
-
     # test 2
-    MOI.set(model, DiffOpt.ForwardConstraintFunction(), c, _vaf(zeros(6)))
+    MOI.set(
+        model,
+        DiffOpt.ForwardConstraintFunction(),
+        c,
+        MOI.VectorAffineFunction{Float64}(
+            MOI.ScalarAffineTerm{Float64}[],
+            zeros(6),
+        ),
+    )
     MOI.set(model, DiffOpt.ForwardObjectiveFunction(), 1.0 * x)
-
     DiffOpt.forward_differentiate!(model)
-
     @test 0.0 ≈ MOI.get(model, DiffOpt.ForwardVariablePrimal(), x) atol = 1e-2 rtol =
         RTOL
+    return
 end
+
+function test_differentiating_lp_with_nonactive_constraints()
+    # Issue #40 from Gurobi.jl
+    # min  x
+    # s.t. x >= 0
+    #      x >= 3
+    qp_test_with_solutions(
+        HiGHS.Optimizer;
+        q = [1.0],
+        G = -ones(2, 1),
+        h = [0.0, -3.0],
+        dzb = -ones(1),
+        dhf = [0.0, 1.0],
+        # Expected solutions
+        z = [3.0],
+        λ = [0.0, 1.0],
+        dzf = -ones(1),
+        dλb = zeros(2),
+        dhb = [0.0, 1.0],
+        ∇zb = zeros(1),
+        ∇λb = [0.0, -1.0],
+        dλf = zeros(2),
+    )
+    return
+end
+
+# min X[1,1] + X[2,2]    max y
+#     X[2,1] = 1         [0   y/2     [ 1  0
+#                         y/2 0    <=   0  1]
+#     X >= 0              y free
+# Optimal solution:
+#
+#     ⎛ 1   1 ⎞
+# X = ⎜       ⎟           y = 2
+#     ⎝ 1   1 ⎠
+function test_differentiating_simple_PSD_back()
+    model = DiffOpt.diff_optimizer(SCS.Optimizer)
+    MOI.set(model, MOI.Silent(), true)
+    X = MOI.add_variables(model, 3)
+    vov = MOI.VectorOfVariables(X)
+    cX = MOI.add_constraint(
+        model,
+        MOI.VectorAffineFunction{Float64}(vov),
+        MOI.PositiveSemidefiniteConeTriangle(2),
+    )
+    c = MOI.add_constraint(
+        model,
+        MOI.VectorAffineFunction(
+            [MOI.VectorAffineTerm(1, MOI.ScalarAffineTerm(1.0, X[2]))],
+            [-1.0],
+        ),
+        MOI.Zeros(1),
+    )
+    MOI.set(
+        model,
+        MOI.ObjectiveFunction{MOI.ScalarAffineFunction{Float64}}(),
+        MOI.ScalarAffineFunction(
+            MOI.ScalarAffineTerm.(1.0, [X[1], X[end]]),
+            0.0,
+        ),
+    )
+    MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
+    MOI.optimize!(model)
+    x = MOI.get(model, MOI.VariablePrimal(), X)
+    cone_types = unique([
+        S for (F, S) in
+        MOI.get(model.optimizer, MOI.ListOfConstraintTypesPresent())
+    ])
+    conic_form = DiffOpt.ConicProgram.Form{Float64}()
+    cones = conic_form.constraints.sets
+    DiffOpt.set_set_types(cones, cone_types)
+    index_map = MOI.copy_to(conic_form, model)
+    @test x ≈ ones(3) atol = ATOL rtol = RTOL
+    MOI.set(model, DiffOpt.ReverseVariablePrimal(), X[1], 1.0)
+    DiffOpt.reverse_differentiate!(model)
+    db = MOI.constant(MOI.get(model, DiffOpt.ReverseConstraintFunction(), c))
+    @test db ≈ [-1.0] atol = ATOL rtol = RTOL
+    return
+end
+
+function test_singular_exception()
+    Q = [4.0 1.0; 1.0 2.0]
+    q = [1.0, 1.0]
+    G = [1.0 1.0]
+    h = [-1.0]
+    model = DiffOpt.diff_optimizer(Ipopt.Optimizer)
+    MOI.set(model, MOI.Silent(), true)
+    x = MOI.add_variables(model, 2)
+    quad_terms = MOI.ScalarQuadraticTerm{Float64}[]
+    for i in 1:2
+        for j in i:2
+            push!(quad_terms, MOI.ScalarQuadraticTerm(Q[i, j], x[i], x[j]))
+        end
+    end
+    objective_function = MOI.ScalarQuadraticFunction(
+        quad_terms,
+        MOI.ScalarAffineTerm.(q, x),
+        0.0,
+    )
+    MOI.set(
+        model,
+        MOI.ObjectiveFunction{MOI.ScalarQuadraticFunction{Float64}}(),
+        objective_function,
+    )
+    MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
+    c = MOI.add_constraint(
+        model,
+        MOI.ScalarAffineFunction(MOI.ScalarAffineTerm.(G[1, :], x), 0.0),
+        MOI.LessThan(h[1]),
+    )
+    MOI.optimize!(model)
+    @test MOI.get(model, MOI.VariablePrimal(), x) ≈ [-0.25; -0.75] atol = ATOL rtol =
+        RTOL
+    @test model.diff === nothing
+    MOI.set.(model, DiffOpt.ReverseVariablePrimal(), x, ones(2))
+    DiffOpt.reverse_differentiate!(model)
+    grad_wrt_h =
+        MOI.constant(MOI.get(model, DiffOpt.ReverseConstraintFunction(), c))
+    @test grad_wrt_h ≈ -1.0 atol = 2ATOL rtol = RTOL
+    @test model.diff !== nothing
+    # adding two variables invalidates the cache
+    y = MOI.add_variables(model, 2)
+    for yi in y
+        MOI.delete(model, yi)
+    end
+    @test model.diff === nothing
+    MOI.optimize!(model)
+    MOI.set.(model, DiffOpt.ReverseVariablePrimal(), x, ones(2))
+    DiffOpt.reverse_differentiate!(model)
+    grad_wrt_h =
+        MOI.constant(MOI.get(model, DiffOpt.ReverseConstraintFunction(), c))
+    @test grad_wrt_h ≈ -1.0 atol = 1e-3
+    @test model.diff !== nothing
+    return
+end
+
+function test_FEASIBILITY_SENSE_zeros_objective()
+    model = DiffOpt.diff_optimizer(HiGHS.Optimizer)
+    MOI.set(model, MOI.Silent(), true)
+    x = MOI.add_variable(model)
+    MOI.add_constraint(model, x, MOI.GreaterThan(1.0))
+    MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
+    MOI.set(model, MOI.ObjectiveFunction{MOI.VariableIndex}(), x)
+    MOI.optimize!(model)
+    @test MOI.get(model, MOI.ObjectiveValue()) ≈ 1.0 atol = ATOL rtol = RTOL
+    MOI.set(model, MOI.ObjectiveSense(), MOI.FEASIBILITY_SENSE)
+    MOI.optimize!(model)
+    @test MOI.get(model, MOI.ObjectiveValue()) == 0.0
+    return
+end
+
+function test_forward_or_reverse_without_optimizing_throws()
+    model = DiffOpt.diff_optimizer(HiGHS.Optimizer)
+    MOI.set(model, MOI.Silent(), true)
+    x = MOI.add_variable(model)
+    MOI.add_constraint(model, x, MOI.GreaterThan(1.0))
+    MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
+    MOI.set(model, MOI.ObjectiveFunction{MOI.VariableIndex}(), x)
+    # do not optimize, just try to differentiate
+    @test_throws ErrorException DiffOpt.forward_differentiate!(model)
+    @test_throws ErrorException DiffOpt.reverse_differentiate!(model)
+    # impossible constraint
+    MOI.add_constraint(model, x, MOI.LessThan(0.5))
+    MOI.optimize!(model)
+    @test_throws ErrorException DiffOpt.forward_differentiate!(model)
+    @test_throws ErrorException DiffOpt.reverse_differentiate!(model)
+    return
+end
+
+struct TestSolver end
+
+# always use IterativeSolvers
+function DiffOpt.QuadraticProgram.solve_system(
+    ::TestSolver,
+    LHS,
+    RHS,
+    iterative::Bool,
+)
+    return IterativeSolvers.lsqr(LHS, RHS)
+end
+
+function test_setting_the_linear_solver_in_the_quadratic_solver()
+    model = DiffOpt.QuadraticProgram.Model()
+    @test MOI.supports(model, DiffOpt.QuadraticProgram.LinearAlgebraSolver())
+    @test MOI.get(model, DiffOpt.QuadraticProgram.LinearAlgebraSolver()) ===
+          nothing
+    MOI.set(model, DiffOpt.QuadraticProgram.LinearAlgebraSolver(), TestSolver())
+    @test MOI.get(model, DiffOpt.QuadraticProgram.LinearAlgebraSolver()) ==
+          TestSolver()
+    MOI.empty!(model)
+    @test MOI.get(model, DiffOpt.QuadraticProgram.LinearAlgebraSolver()) ==
+          TestSolver()
+    return
+end
+
+function test_quad_to_soc()
+    model = DiffOpt.ConicProgram.Model()
+    F = MOI.VectorAffineFunction{Float64}
+    S = MOI.RotatedSecondOrderCone
+    # Adds the set type to the list
+    @test MOI.supports_constraint(model, F, S)
+    bridged = MOI.Bridges.Constraint.QuadtoSOC{Float64}(model)
+    x = MOI.add_variables(bridged, 3)
+    # We pick orthogonal rows but this is not the `U` that cholesky is going to
+    # take in the bridge.
+    U = [1 -1 1; 1 2 1; 1 0 -1.0]
+    a = [1, 2, -3.0]
+    b = -5.0
+    Q = U' * U
+    f = x' * Q * x / 2.0 + a' * x
+    c = MOI.add_constraint(bridged, f, MOI.LessThan(-b))
+    dQ = [1 -1 0; -1 2 1; 0 1 3.0]
+    da = [-1, 1.0, -2]
+    db = 3.0
+    df = x' * dQ * x / 2.0 + da' * x + db
+    MOI.Utilities.final_touch(bridged, nothing)
+    MOI.set(bridged, DiffOpt.ForwardConstraintFunction(), c, df)
+    return
+end
+
+function _test_dU_from_dQ(U, dU)
+    dQ = dU'U + U'dU
+    _dU = copy(dQ)
+    __dU = copy(dQ)
+    # Compiling
+    DiffOpt.dU_from_dQ!(__dU, U)
+    @test @allocated(DiffOpt.dU_from_dQ!(_dU, U)) == 0
+    @test _dU ≈ dU
+    return
+end
+
+function test_dU_from_dQ()
+    _test_dU_from_dQ(2ones(1, 1), 3ones(1, 1))
+    U = [1 2; 0 1]
+    dU = [1 -1; 0 1]
+    _test_dU_from_dQ(U, dU)
+    U = [-3 5; 0 2.5]
+    dU = [2 3.5; 0 -2]
+    _test_dU_from_dQ(U, dU)
+    U = [1.5 2 -1; 0 -1 3.5; 0 0 -2]
+    dU = [2.5 -1 2; 0 5 -3; 0 0 3]
+    _test_dU_from_dQ(U, dU)
+    return
+end
+
+end  # module
+
+TestMOIWrapper.runtests()
