+++ test/runtests.jl
@@ -4,77 +4,6 @@
 # in the LICENSE.md file or at https://opensource.org/licenses/MIT.
 
 using Test
-import DiffOpt
-import Random
-import LinearAlgebra
-import DelimitedFiles
-import MathOptInterface as MOI
-const MOIT = MOI.Test
 
-const ATOL = 2e-4
-const RTOL = 2e-4
-
-include("bridges.jl")
-
-@testset "MOI_wrapper" begin
-    @testset "Utils" begin
-        include("utils.jl")
-    end
-    @testset "MOI_wrapper main" begin
-        include("moi_wrapper.jl")
-    end
-    @testset "QP fwd" begin
-        include("qp_forward.jl")
-    end
-    @testset "Conic reverse" begin
-        include("conic_reverse.jl")
-    end
-end
-
-@testset "JuMP wrapper" begin
-    include("jump.jl")
-end
-
-@testset "Solver Interface" begin
-    include("solver_interface.jl")
-end
-
-@testset "Singular error with deleted variables: Sensitivity index issue" begin
-    include("singular_exception.jl")
-end
-
-@testset "Examples" begin
-    @testset "autotuning-ridge" begin
-        include(joinpath(@__DIR__, "../docs/src/examples/autotuning-ridge.jl"))
-    end
-    @testset "chainrules_unit" begin
-        include(joinpath(@__DIR__, "../docs/src/examples/chainrules_unit.jl"))
-    end
-    @testset "custom-relu" begin
-        include(joinpath(@__DIR__, "../docs/src/examples/custom-relu.jl")) # needs downloads
-    end
-    @testset "matrix-inversion-manual" begin
-        include(
-            joinpath(
-                @__DIR__,
-                "../docs/src/examples/matrix-inversion-manual.jl",
-            ),
-        )
-    end
-    @testset "sensitivity-analysis-ridge" begin
-        include(
-            joinpath(
-                @__DIR__,
-                "../docs/src/examples/sensitivity-analysis-ridge.jl",
-            ),
-        )
-    end
-    @testset "sensitivity-analysis-svm" begin
-        include(
-            joinpath(
-                @__DIR__,
-                "../docs/src/examples/sensitivity-analysis-svm.jl",
-            ),
-        )
-    end
-end
+include("moi_wrapper.jl")
+include("jump.jl")
